move the gallery tests to under scaffolding
